Prevent Heroku app from sleeping
Heroku normally set an app in sleeping mode when there is no activity.
The disadventage for our app is that the next request takes more time
to load. This is a bad experience for visitors. We prevent from going
to sleep by making sure that there is activity within every minute.